Add files via upload
The task was to create a Smart Meter and simulate the activity in a house. More details will be found in the ReadMe text file. Here is the coursework:
 The aim of this coursework was to construct a simple simulation of the energy consumption of a house. This house will have a number of smart meters in it that record the amount of gas, water and electricity being
used. The house will contain a number of appliances that when operating will consume varying amounts of
the resources. People in the house will turn these appliances on and off at different times.